requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
arrow==1.3.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
blessed==1.20.0
Brotli==1.1.0
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
dj-database-url==2.3.0
Django==5.1.4
django-picklefield==3.2
django-rq==3.0.0
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
requests==2.32.3
rq==2.0.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.2
tenacity==9.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.18.3
